requirements.txt
django
django-compressor
django-templatetag-handlebars
django-easy-pjax
djangorestframework

-e git+https://github.com/laginha/django-mobileesp/#egg=django_mobileesp
